Move markers to types
we keep them undocumented because they are impl detail
